+++ sparkwing/inspect_contract_test.go
@@ -1,0 +1,9 @@
+package sparkwing
+
+import (
+	"context"
+
+	"github.com/sparkwing-dev/sparkwing/pkg/pipelines"
+)
+
+var _ func(context.Context, *Registration, *pipelines.Pipeline) ([]SecretField, error) = InspectPipelineSecrets